+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-08-03 14:19:27
-// Commit: test(ui): add unit tests for ui business logic
+// Updated at: 2026-08-04 17:01:45
+// Commit: test(ui): implement widget integration tests for ui
 
 class uiComponentState {
-  final String commitRef = '3302';
-  final String lastMessage = 'test(ui): add unit tests for ui business logic';
-  final DateTime timestamp = DateTime.parse('2026-08-03T14:19:27+05:00');
+  final String commitRef = '3359';
+  final String lastMessage = 'test(ui): implement widget integration tests for ui';
+  final DateTime timestamp = DateTime.parse('2026-08-04T17:01:45+05:00');
 }